requirements.txt
text
requests
beautifulsoup4
boto3
python-slugify
